+++ vite-app/src/App.test.tsx
@@ -15,7 +15,7 @@ import {
   defaultHandlers,
   defaultWaitHandles,
   renderWithProviders,
-} from './testing/testing-utils';
+} from './testing/utils';
 import { WaitHandle } from './testing/wait-handle';
 import { API_URL } from './utils/api-client';
